requirements.txt
flask==2.3.2
werkzeug==2.3.6
openai==0.27.8
python-docx==0.8.11
PyMuPDF==1.21.1
python-dotenv==1.0.0